requirements.txt
boto3
google-cloud-secret-manager
PyYAML
